Initial
Adds more content then should exist in one commit. The important thing to note is the main idea of what I'm doing for the rebase.

Site-104 was made for F19 (Tegu Code). Many paths do not exist on Daed. So I am replacing all examples of unrecognized paths with compatible versions on Daed, marking them for porting and setting aside later, or outright removing them from my map.

The goal is to retain as much of the decoration and soul of my map as possible, as I don't quite fancy redoing so much work.

Ported Flaregun, and added some logic for shooting a flare up into the sky in order to attempt to get a rescue ship/aid.